demande la confirmation au moment de supprimer la liste des favoris

+++ script/favorites.js
@@ -1,7 +1,11 @@
 $(function() {
     $("#clear-button").on('click',function() {
-        console.log(localStorage);
-        localStorage.clear();
-        console.log(localStorage);
+        if (confirm("Voulez vous vraiment supprimer tous vos favoris ?")) {
+            localStorage.clear();
+            console.log(localStorage);
+        } else {
+            console.log(localStorage);
+        }
+        
     });
 });
